requirements.txt
dateutils
django-mptt>=0.2.1
easy_thumbnails>=1.0-alpha-17
django-polymorphic>=0.2
django-filer
django-taggit

